Moved RelatedFieldHandler functionality to BaseRelatedFieldHandler
RelatedFieldHandler functionality was used not only by simple field
handlers, but also as BaseFileFieldHandler, so it makes more sense
to create a base class out of it.
